refactor: removes Payments namespace
- also moves payments.d.ts to /common/types
