+++ src/features/user/Account/components/AccountRecord.tsx
@@ -8,11 +8,16 @@ import { TFunction } from 'next-i18next';
 import DownloadCodes from './DownloadCodes';
 import BackButton from '../../../../../public/assets/images/icons/BackButton';
 import TransferDetails from './TransferDetails';
+import {
+  PaymentDetails,
+  PaymentHistoryRecord,
+  RecipientBank,
+} from '../../../common/types/payments';
 
 interface CommonProps {
   handleRecordToggle: (index: number | undefined) => void;
   selectedRecord: number | null;
-  record: Payments.PaymentHistoryRecord;
+  record: PaymentHistoryRecord;
   isPlanetCash?: boolean;
 }
 
@@ -99,7 +104,7 @@ export default function AccountRecord({
 }
 
 interface HeaderProps {
-  record: Payments.PaymentHistoryRecord;
+  record: PaymentHistoryRecord;
   handleRecordToggle?: (index: number | undefined) => void;
   index?: number;
   isPlanetCash?: boolean;
@@ -177,7 +182,7 @@ export function RecordHeader({
 }
 
 interface DetailProps {
-  record: Payments.PaymentHistoryRecord;
+  record: PaymentHistoryRecord;
 }
 
 export function DetailsComponent({ record }: DetailProps): ReactElement {
@@ -366,7 +371,7 @@ export function DetailsComponent({ record }: DetailProps): ReactElement {
 }
 
 export const showStatusNote = (
-  record: Payments.PaymentHistoryRecord,
+  record: PaymentHistoryRecord,
   t: TFunction
 ): ReactElement => {
   const showDonationNote = (): string => {
@@ -392,7 +397,7 @@ export const showStatusNote = (
 };
 
 interface BankDetailsProps {
-  recipientBank: Payments.RecipientBank;
+  recipientBank: RecipientBank;
 }
 
 export function BankDetails({ recipientBank }: BankDetailsProps): ReactElement {
@@ -458,7 +463,7 @@ export function BankDetails({ recipientBank }: BankDetailsProps): ReactElement {
 }
 
 interface CertificatesProps {
-  recordDetails: Payments.PaymentDetails;
+  recordDetails: PaymentDetails;
 }
 
 export function Certificates({
